[refac] : Test Case Refactoring
