+++ src/main/java/com/example/fastboard/domain/member/controller/MemberController.java
@@ -8,7 +8,6 @@ import com.example.fastboard.domain.member.service.MemberSaveService;
 import com.example.fastboard.global.common.response.ApiResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
