+++ src/main/java/com/example/fastboard/domain/member/dto/request/MemberSaveReq.java
@@ -9,13 +9,6 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 public class MemberSaveReq {
-    /**
-     * TODO : DTO Validation 필요.
-     * email : NOT NULL
-     * name : NOT NULL,
-     * password : NOT NULL.
-     */
-
     @Email(message = "이메일 형식을 지켜주세요.")
     @NotNull(message = "이메일을 입력해주세요.")
     private String email;
@@ -23,10 +16,13 @@ public class MemberSaveReq {
     @NotBlank(message = "이름을 입력해주세요.")
     @NotNull(message = "이름을 입력해주세요.")
     private String name;
+
     @NotBlank(message = "비밀번호를 입력해주세요.")
     @NotNull(message = "비밀번호를 입력해주세요.")
     private String password;
+
     private String nickname;
+
     private String phoneNumber;
 
 }
